feat: Объекдинить пакеты stats и fetch Удалить старую логику отправки спрятать логику работы PollCount
