+++ internal/report/fetch_test.go
@@ -1,4 +1,4 @@
-package stats
+package report
 
 import (
 	"testing"
