+++ internal/report/names.go
@@ -8,7 +8,7 @@
 //	DropPoll - сбрасывает значение PollCount
 //
 // В целом этот пакет смотрит достаточно абстратно на метрики, просто сохраняет их в хранилище, дальше ему все равно
-package stats
+package report
 
 const (
 	// название рамндомной метрики среди всех данных, что мы собираем
