+++ internal/report/send_test.go
@@ -1,8 +1,12 @@
 package report_test
 
 import (
+	"bytes"
+	"io"
+	"net/http"
 	"net/http/httptest"
 	"reflect"
+	"regexp"
 	"testing"
 
 	"github.com/mailru/easyjson"
@@ -77,4 +81,62 @@ func wrapFloat64(v int) (ref *float64) {
 	ref = new(float64)
 	*ref = float64(v)
 	return
+}
+
+// testHandler простая структура, которая запоминает по каким адресам к ней делали запросы
+// используется в тестировании.
+type testHandler struct {
+	requests []*http.Request
+}
+
+// ServerHTTP отвечает на запросы к серверы, исполяет интерфейс http.Handle
+func (server *testHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
+	bb, _ := io.ReadAll(r.Body)
+	defer r.Body.Close()
+	// чтобы потом можно было прочитать тело запроса,
+	// мы его сохраняем в сцециализированную переменную внутри запроса,
+	// которая работает как замыкаение
+	r.GetBody = func() (io.ReadCloser, error) {
+		return io.NopCloser(bytes.NewBuffer(bb)), nil
+	}
+
+	// добавляем запросы в массив запросов. Теперь каждый такой запрос помнит и тело своего запроса
+	server.requests = append(server.requests, r)
+}
+
+// routesUsed возвращает марштуры по которым проходили запросы
+func (server testHandler) routesUsed() (routes []string) {
+	for _, r := range server.requests {
+		routes = append(routes, r.URL.Path)
+	}
+	return
+}
+
+// containsRoute возвращает true, если в принятых сервером маршрутах находится такой.
+// Задается регулярным вырежаением pattern
+func (server testHandler) containsRoute(pattern string) (bool, error) {
+	for _, r := range server.requests {
+		found, err := regexp.MatchString(pattern, r.URL.Path)
+		if err != nil {
+			return false, err
+		}
+		if found {
+			return true, nil
+		}
+	}
+	return false, nil
+}
+
+// возвращает количество полученных запросов
+func (server testHandler) NumRequests() int {
+	return len(server.requests)
+}
+
+// stringerWrap позволяет использовать интерфейс стрингер в полях структуры. Позволяя запихивать туда люую переменную, которую можно обратить в строку
+type stringerWrap struct {
+	text string
+}
+
+func (s stringerWrap) String() string {
+	return s.text
 }
